Authentication is now performed via its dedicated looper. This helps to prevent unnecessary copies of the password in memory.
Disabled remote (not in-app) scripting and drag-and-drop
messages to prevent revealing the password through it
(for example: hey get Value of View "tc_password" of View ...).

Secure events, logging events and shutdown operations are now
performed at BApplication level and not in BWindow.

Fixed incorrect iterations of image paths during image list
or image folder path filtering.

SessionBar buttons are now not flat. Because they do not contain
icons, formerly they could be missed out as a text.